Cleaned up Panache.ts errors and fixed a bug with the demoralize immunity removal always triggering.

+++ src/effects/panache.ts
@@ -1,24 +1,35 @@
-import { ChatMessagePF2e } from "foundry-pf2e";
+import { ActorPF2e, ChatMessagePF2e } from "foundry-pf2e";
+import { logd } from "../utils.ts";
 
-export function checkForBravado(chatMessage) {
-
+export function checkForBravado(chatMessage: ChatMessagePF2e) {
+  
 //don't run if tumble through or enjoy the show - those hooks will call this function after the animation
-if (chatMessage.flags?.pf2e?.context?.options.includes("item:trait:bravado")
-  && !chatMessage.flags?.pf2e?.context?.options.includes("action:tumble-through")
-  && !chatMessage.flags?.pf2e?.context?.options.includes("item:slug:enjoy-the-show")) {
+if (chatMessage.flags?.pf2e?.context?.options?.includes("item:trait:bravado")
+  && !chatMessage.flags?.pf2e?.context?.options?.includes("action:tumble-through")
+  && !chatMessage.flags?.pf2e?.context?.options?.includes("item:slug:enjoy-the-show")) {
     checkIfProvidesPanache(chatMessage);
   }
 }
   
-export async function checkIfProvidesPanache(chatMessage) {
-  const outcome = chatMessage.flags.pf2e.context.outcome;
-  if (chatMessage.flags?.pf2e?.context?.options.includes("item:trait:bravado") 
+export async function checkIfProvidesPanache(chatMessage: ChatMessagePF2e) {
+  const outcome = chatMessage.flags?.pf2e?.context?.outcome;
+  if (chatMessage.flags?.pf2e?.context?.options?.includes("item:trait:bravado") 
     && (outcome === "criticalSuccess" || outcome === "success" || outcome === "failure")) {
-    await applyPanache(game.actors.get(chatMessage.speaker.actor), outcome);
+    const actorId = chatMessage.speaker.actor;
+    if (!actorId) {
+      logd("Actor ID not found in chat message.");
+      return;
+    }
+    const actor = game.actors.get(actorId);
+    if (!actor) {
+      logd("Cannot find actor with ID: " + actorId + ".");
+      return;
+    }
+  await applyPanache(actor, outcome);
   }
 }
 
-async function applyPanache(actor, outcome) {
+async function applyPanache(actor: ActorPF2e, outcome: "success" | "failure" | "criticalSuccess") {
   const existingPanacheStatus = await hasPanache(actor);
 
   if (existingPanacheStatus === "success") {
@@ -29,8 +40,16 @@ async function applyPanache(actor, outcome) {
     }
   } else {
     const panacheItemId = "uBJsxCzNhje8m8jj";
-    const compendiumPack = game.packs.get("pf2e.feat-effects");
+    const compendiumPack = game?.packs?.get("pf2e.feat-effects");
+    if (!compendiumPack) {
+      logd("Compendium not found in Game.");
+      return;
+    }
     const panacheEffect = await compendiumPack.getDocument(panacheItemId);
+    if (!panacheEffect){
+      logd("Panache effect not found in Compendium pack.");
+      return;
+    }
     await actor.createEmbeddedDocuments("Item", [panacheEffect.toObject()]);
 
     if (outcome === "failure") {
@@ -39,7 +58,7 @@ async function applyPanache(actor, outcome) {
   }
 }
 
-async function hasPanache(actor) {
+async function hasPanache(actor: ActorPF2e) {
   const items = await actor.items.contents;
   const panacheEffect = items.find(item =>
     item.type === "effect" && item.system.slug === "effect-panache"
@@ -50,11 +69,15 @@ async function hasPanache(actor) {
     : false;
 }
 
-async function editPanacheEffect(actor, outcome) {
+async function editPanacheEffect(actor: ActorPF2e, outcome: "success" | "failure" | "criticalSuccess") {
   try {
     const panacheEffect = actor.items.find(item =>
       item.type === "effect" && item.system.slug === "effect-panache"
     );
+
+    if (!panacheEffect) {
+      throw new Error("Panache effect not found.");
+    }
 
     const updatedSuccessEffectData = {
       name: "Effect: Panache",
@@ -75,32 +98,41 @@ async function editPanacheEffect(actor, outcome) {
     }
 
   } catch (error) {
-    ui.notifications.error(`Error updating Panache effect: ${error.message}`);
+    ui.notifications.error(`Error updating Panache effect: ${(error as Error).message}`);
     console.error(error);
   }
 }
 
-export async function checkForFinisherAttack(chatMessage) {
+export async function checkForFinisherAttack(chatMessage: ChatMessagePF2e) {
   const context = chatMessage.flags?.pf2e?.context;
-  if (!context?.options.includes("finisher")) {
+  if (context === undefined || !context.options?.includes("finisher")) {
     return;
   }
   if (context?.outcome === "failure" || context?.outcome === "criticalFailure") {
+    //add logic for button to remove panache
     clearPanache(chatMessage);
+  } else {
+    removeDemoralizeImmunity(chatMessage);
   }
-  removeDemoralizeImmunity(chatMessage);
-
 }
 
-export async function checkForFinisherDamage(chatMessage) {
-  if (!chatMessage.flags?.pf2e?.context?.options.includes("finisher")) {
+export async function checkForFinisherDamage(chatMessage: ChatMessagePF2e) {
+  if (!chatMessage.flags?.pf2e?.context?.options?.includes("finisher")) {
     return;
   }
   clearPanache(chatMessage);
 }
 
-function clearPanache(chatMessage) {
-  const panacheItems = returnPanacheItems(game.actors.get(chatMessage.speaker.actor));
+function clearPanache(chatMessage: ChatMessagePF2e) {
+  const actorId = chatMessage.speaker.actor;
+  if (!actorId) {
+    return;
+  }
+  const actor = game.actors.get(actorId);
+  if (!actor) {
+    return;
+  }
+  const panacheItems = returnPanacheItems(actor);
   if (panacheItems.length > 0) {
     for (const panacheItem of panacheItems) {
       panacheItem.delete();
@@ -108,12 +140,12 @@ function clearPanache(chatMessage) {
   }
 }
 
-function returnPanacheItems(actor) {
+function returnPanacheItems(actor: ActorPF2e) {
   const items = actor.items.contents;
   return items.filter(item => item.type === "effect" && item.system.slug === "effect-panache");
 }
 
-export async function checkForExtravagantParryOrElegantBuckler(chatMessage) {
+export async function checkForExtravagantParryOrElegantBuckler(chatMessage: ChatMessagePF2e) {
   const { flags } = chatMessage;
 
   if (!flags?.pf2e?.context || flags.pf2e.context.type !== "attack-roll" || !flags.pf2e.context.target?.actor) {
@@ -121,7 +153,14 @@ export async function checkForExtravagantParryOrElegantBuckler(chatMessage) {
   }
 
   const targetActorId = flags.pf2e.context.target.actor.split('.').pop();
+  if (!targetActorId) {
+    return;
+  }
   const target = game.actors.get(targetActorId);
+
+  if (!target) {
+    return;
+  }
 
   const hasDuelingParry = flags.pf2e.context.options.includes("target:effect:dueling-parry");
   const hasShieldRaised = flags.pf2e.context.options.includes("target:effect:raise-a-shield");
@@ -138,18 +177,20 @@ export async function checkForExtravagantParryOrElegantBuckler(chatMessage) {
   }
 }
 
-async function hasElegantBucklerFeat(actor) {
+async function hasElegantBucklerFeat(actor: ActorPF2e) {
   const items = await actor.items.contents;
   return items.some(item => 
-    item.type === "feat" && 
+    item.type === "feat" &&
+    // @ts-expect-error - category does exist on item type "feat" 
     item.system.category === "class" && 
     item.system.slug === "elegant-buckler"
   );
 }
 
 async function removeDemoralizeImmunity(chatMessage: ChatMessagePF2e) {
-    const { context } = chatMessage.flags.pf2e ?? {};
-    const attacker = game.actors.get(chatMessage.speaker.actor);
+  const { context } = chatMessage.flags.pf2e ?? {};
+    const attackerActorId = chatMessage.speaker.actor;
+    const attacker = game.actors.get(attackerActorId as string);
     const target = chatMessage.target?.actor;
     const contextOptions = new Set(context?.options ?? []);
